+++ src/app/form.config.ts
@@ -1,4 +1,4 @@
-import { FormFieldConfig } from './models/models';
+import { FormFieldConfig } from './shared/models/models';
 
 export const AUDIO_TRACK_FORM_CONFIG: FormFieldConfig[] = [
   {