+++ src/app/shared/models/models.ts
@@ -1,0 +1,9 @@
+export interface FormFieldConfig {
+  label: string;
+  inputType: string;
+  inputName: string;
+  inputPlaceholder: string;
+  formControlName: string;
+}
+
+export type Playlist = string[];